add postprocessing algorithm for final submission

+++ DEEP-PSMA/DEEP-PSMA_Refine.py
@@ -10,7 +10,7 @@ def expand_contract_label(label,distance=5.):
     distance_filter.SquaredDistanceOff() #default distances computed as d^2, set to linear distance
     dmap=distance_filter.Execute(label) #execute on label, returns SITK image in same space with voxels as distance from surface
     dmap_ar=sitk.GetArrayFromImage(dmap) #array of above SITK image
-    new_label_ar=(dmap_ar<=distance).astype('int16') #binary array of values less than or equal to selected d
+    new_label_ar=(dmap_ar<=distance).astype('int8') #binary array of values less than or equal to selected d
     new_label=sitk.GetImageFromArray(new_label_ar) #create new SITK image and copy spatial information
     new_label.CopyInformation(label)
     return new_label
@@ -48,7 +48,7 @@ def refine_my_ttb_label(ttb_image,pet_image,totseg_multilabel,expansion_radius_m
     if normal_image is not None:
         normal_array=sitk.GetArrayFromImage(normal_image)
         ttb_rethresholded_array[normal_array>0]=0
-    ttb_rethresholded_image=sitk.GetImageFromArray(ttb_rethresholded_array.astype('int16')) #create output image & copy information
+    ttb_rethresholded_image=sitk.GetImageFromArray(ttb_rethresholded_array.astype('int8')) #create output image & copy information
     ttb_rethresholded_image.CopyInformation(ttb_image)
     return ttb_rethresholded_image
